+++ src/protocol.c
@@ -209,9 +209,9 @@ uint8_t temperature_value_to_degrees_f(uint8_t val)
 /**
  * Converts kilopascals into pounds per square inch.
  */
-uint8_t kpa_to_psi(uint8_t kpa)
+float kpa_to_psi(uint8_t kpa)
 {
-    return (uint8_t)(kpa / 6.89475729);
+    return ((float)kpa / 6.89475729);
 }
 
 /**
